# openapi.yaml
...
openapi: 3.1.0
info:
  title: Nexus API
  version: 0.1.0
  description: "Generated from the live Fastify route table. Every path carries a
    maturity tag: `durable` — Dedicated handler over durable storage. Safe to
    build on. `dedicated-volatile` — Dedicated handler, but its state lives in
    process memory and is lost on restart. `bridge` — Served by the legacy
    in-memory bridge (routes/api-bridge.ts). Treat its data as synthetic."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/discussion/stream:
    post:
      tags:
        - durable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - topic
                - participants
              properties:
                topic:
                  type: string
                  minLength: 1
                  maxLength: 10000
                participants:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - label
                      - provider
                      - model
                    properties:
                      label:
                        type: string
                        minLength: 1
                      provider:
                        type: string
                        minLength: 1
                      model:
                        type: string
                        minLength: 1
                supervisorModel:
                  type: string
                settlePauseMs:
                  type: number
                  minimum: 0
                maxContributions:
                  type: number
                  minimum: 1
                maxWallMs:
                  type: number
                  minimum: 1000
      responses:
        "200":
          description: Default Response
...
  /api/v1/governance/approvals/{approvalId}:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: approvalId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/governance/approvals/{approvalId}/approve:
    post:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: approvalId
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/governance/approvals/{approvalId}/reject:
    post:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: approvalId
          required: true
      responses:
        "200":
          description: Default Response
...
security:
  - bearerAuth: []
...